broadphase: fix bug in nodeCollide
When one node is a leaf and a plane/halfspace, if the other node
contains a plane/halfspace, the overlap test returns a wrong result (it
uses a support point computation but finds it at infinity due to the
presence of an infinite AABB). Fix: we should simply check for overlap
of the aabbs of node1 and node2.